Commit adds new commands to the PARAM.in file for debugging and controlling for where the plasmaphere is currently defined in ModWaveDiffusion:
(1) Rather than editing the source code to print out and investigate
individual field lines, the command #TESTFIELDLINE accepts integer
Latidude and Longitude indices for the field line to be output.

(2) Added command #PLASMAPAUSEDENSITY to specify the value of
DensityP, where the different wave models are either applied inside or
outside of the plasmasphere for the present MLT.  This fix does NOT
change the calculation to be gradient-based, but has been tested in
applications of the Simple (Kp-driven) and SAMI3 plasmasphere models.