Commit adds specifying the output grid at energies for RBSP MagEIS and REPT instruments.  Still working on this since the scaling of mu and K grids are currently hard-coded in with "magic numbers."  Will test sensitivity for:
(1) oversampling the mu grid for larger neng in the output energy
grid,

and

(2) fixing the "magic numbers" that determine the mu and K grids so
that arbitrary energy grids can fit into the mu-K space.

Also, neng is now an arbitrary number which can be specified by the
user in the PARAM.in file.  As a result variables in Natasha's
BoundaryCheck module had to move to ModCimi.f90 for allocation.