Commit completing changes made to xmm grid initialization such that the highest energy particles will fit in xmm at 100 R_E well beyond the outer boundary of the simulation grid.  Lowest energy particles will fit in xmm at L=1. No changes made to the K-grid at this time.
Also moved rb variable from ModFieldTrace to the ModGrid.  Largely
unnecessary variable with associated logical variable UseExpandedGrid
where the logical is only checked in ModFieldTrace and then rb is set.
This variable is directly set in the grid file to begin with, and can
be further set in the PARAM.in file.

Output routine for PSD updated to make it clearer on the conversion
factors of xmm and xk grids.