Makes changes to the output file names and some contents of the files.
Log File:
CIMI now outputs the logfile which contains the iteration number.
This file is appended with successive writes.  Upon restart, a new
logfile is created with the new iteration number.  THese files will
need to be concatenated with post-processing routines.

All Files:
If SaveSeparate files is set true for a file type, correctly outputs
the current time up to second precision (HHMMSS) depending on the
output cadence selected in the PARAM file.

Satellite output:
Lists the number of energy and Pitch Angles in the file header for
easy looping to get pitch angle resolved fluxes or for calculating
omnidirectional flux.