Fixed some compilation issues in ModDiagDiff while using the nagfor compiler.  Converted several large arrays to allocatable and added another subroutine to allocate these arrays.
***NOTE***
ModDiagDiff is not working correctly at the moment using the default
input/testfiles/PARAM.in.test.DiagDiff.  The DiagDiff code will not
work even without the changes committed here.

Added a broadcast subroutine of ModFieldTrace variables to broadcast
bo, bm, volume, etc. to all processors.  (All variables in the gather
subroutine can be broadcast to all PEs if needed.)  This was added to
assist Lstar calculation, but could be helpful in other ways for
future development.